Fix assertion failure when starting the About box (format string has wrong type).

+++ src/hugin1/hugin/AboutDialog.cpp
@@ -133,10 +133,12 @@ void AboutDialog::GetSystemInformation(wxFont *font)
     else
         is64=_("32 bit");
     text=text+wxT("\n")+wxString::Format(_("Architecture: %s"),is64.c_str());
-    text=text+wxT("\n")+wxString::Format(_("Free memory: %d kiB"),wxGetFreeMemory().GetValue()/1024);
+    // wxGetFreeMemory returns a wxMemorySize, which is undocumented.
+    // However, we know -1 is returned on failure, so it must be signed.
+    text=text+wxT("\n")+wxString::Format(_("Free memory: %ld kiB"),(long int) wxGetFreeMemory().GetValue()/1024);
 #ifdef _WINDOWS
     UINT cp=GetACP();
-    text=text+wxT("\n")+wxString::Format(_("Active Codepage: %d"),cp); 
+    text=text+wxT("\n")+wxString::Format(_("Active Codepage: %u"),cp); 
     switch(cp)
     {
     case 1250:
